twig: Replace 'index' with 'tree' command
