+++ src/bin/twig/cli.rs
@@ -7,8 +7,8 @@ use tooling::{
 };
 
 pub mod common;
-mod index;
 mod odb;
+mod tree;
 
 #[derive(Parser)]
 pub struct Cli {
@@ -29,8 +29,8 @@ pub struct Cli {
 pub enum TwigCommand {
     /// Perform operations on or with the object database
     Odb(odb::CommandOdb),
-    /// Work with or create index files
-    Index(index::CommandIndex),
+    /// Work with or create trees
+    Tree(tree::CommandTree),
 }
 
 impl Cli {
@@ -71,7 +71,7 @@ impl TwigCommand {
     pub fn run(&self, cli: &Cli) -> Result<i32, Error> {
         match self {
             Self::Odb(cmd) => cmd.run(cli),
-            Self::Index(cmd) => cmd.run(cli),
+            Self::Tree(cmd) => cmd.run(cli),
         }
     }
 }
